+++ balm/drivers.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from biodivine_aeon import BooleanNetwork
 
-from nfvsmotifs.space_utils import percolate_space
+from balm.space_utils import percolate_space
 
 
 def find_single_node_LDOIs(bn: BooleanNetwork) -> dict[tuple[str, int], dict[str, int]]:
