+++ tests/petri_net_translation_test.py
@@ -3,7 +3,7 @@ import pytest
 from biodivine_aeon import BooleanNetwork, RegulatoryGraph  # type: ignore
 from networkx import DiGraph, is_isomorphic  # type: ignore
 
-from nfvsmotifs.petri_net_translation import (
+from balm.petri_net_translation import (
     extract_variable_names,
     network_to_petrinet,
     sanitize_network_names,
